Fix before-commit lease isolation spec
